+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   `text[]` (known as `_text` in the `pg_type` system table), `int4[]`, `float4[]`, `float8[]` can now
   be passed as function arguments.
 
+- Add serialization support for the `_varchar` data type.
+
 - Add workarounds and detection code for the CedarDB PostgreSQL variant.
 
 - New error classes `pg-duplicate-table` and `pg-duplicate-column`, subclasses of